+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * **x.X versions** are either updates that add content or major bug fixes
 * **x.x.X versions** are either small content update (language translation, keybind change, ...) or bug fixes
 
-## v1.6.0
+## v?.?.?
 
 * Added volume increment feature ([@naturecodevoid](https://github.com/naturecodevoid))
 * Added feature that allows user to increment the music volume up to 200% (disabled by default) ([@naturecodevoid](https://github.com/naturecodevoid))